[re-factory] clean editor components
